+++ documentation/js/menu-wc.js
@@ -240,6 +240,12 @@ customElements.define('compodoc-menu', class extends HTMLElement {
                         <a href="interfaces/ComponentMockOption.html" data-type="entity-link">ComponentMockOption</a>
                     </li>
                     <li class="link">
+                        <a href="interfaces/Device.html" data-type="entity-link">Device</a>
+                    </li>
+                    <li class="link">
+                        <a href="interfaces/Group.html" data-type="entity-link">Group</a>
+                    </li>
+                    <li class="link">
                         <a href="interfaces/Notification.html" data-type="entity-link">Notification</a>
                     </li>
             </ul>