+++ src/app/add-devices-group/add-devices-group.component.ts
@@ -5,7 +5,6 @@ import { BackendService } from '../services/backend.service';
 import { MockupBackendService } from '../services/mockup-backend.service';
 import { NotificationsService } from '../services/notifications.service';
 import { LocalStorageKeys } from '../definitions/const/local-storage-keys';
-import { FormGroup } from '@angular/forms';
 import { DeviceGroupCreatedComponent } from '../device-group-created/device-group-created.component';
 
 @Component({
@@ -14,24 +13,27 @@ import { DeviceGroupCreatedComponent } from '../device-group-created/device-grou
   styleUrls: ['./add-devices-group.component.scss']
 })
 export class AddDevicesGroupComponent implements OnInit {
-  checkBox = true;
   /**
    * Backend reference
    */
   backend: Backend;
 
   /**
-   * Models that hold organization id, user role, name and email
+   * Models that hold organization id
    */
   organizationId: string;
-  userRole: string;
-  userName: string;
-  email: string;
-  password: string;
-  passwordConfirm: string;
-
+  defaultConnectivity: boolean;
+  groupApiKey: string;
+  enabled: boolean;
   errorMessages: string[];
 
+  /**
+   *  Models that hold group data
+   */
+  device_group_id: string;
+  name: string;
+  default_device_connectivity: boolean;
+  device_group_api_key: string;
 
   /**
    * Models that removes the possibility for the user to close the modal by clicking outside the content card
@@ -41,11 +43,6 @@ export class AddDevicesGroupComponent implements OnInit {
     ignoreBackdropClick: true
   };
 
-  /**
-   * Device grup modal window reference
-   */
-  bsDeviceGroupModalRef: BsModalRef;
-
   constructor(
     public bsModalRef: BsModalRef,
     private modalService: BsModalService,
@@ -53,8 +50,10 @@ export class AddDevicesGroupComponent implements OnInit {
     private mockupBackendService: MockupBackendService,
     private notificationsService: NotificationsService
   ) {
-    this.userRole = null;
-    const mock = localStorage.getItem(LocalStorageKeys.addUserMock) || null;
+    this.defaultConnectivity = false;
+    this.enabled = false;
+    this.groupApiKey = 'Loading...';
+    const mock = localStorage.getItem(LocalStorageKeys.addGroupMock) || null;
     // check which backend is required (fake or real)
     if (mock && mock === 'true') {
       this.backend = mockupBackendService;
@@ -62,6 +61,11 @@ export class AddDevicesGroupComponent implements OnInit {
       this.backend = backendService;
     }
     this.errorMessages = [];
+    this.device_group_id = 'Loading ...';
+    this.name = 'Loading ...';
+    this.enabled = false;
+    this.default_device_connectivity = false;
+    this.device_group_api_key = 'Loading ...';
 
   }
 
@@ -69,36 +73,34 @@ export class AddDevicesGroupComponent implements OnInit {
   }
 
   /**
-   * Requests to add a new user
-   * @param form Form with the user input data
+   * Requests to create a new device group
+   * Create device group opens the device group created component modal  window confirmation
+   * @param form Form with the group input data
    */
-  addUser(form) {
+  addGroup(form) {
     if (this.errorMessages.length === 0) {
-      const user = {
+      const groupData = {
         name: form.value.name,
-        email: form.value.email,
-        password: form.value.password,
-        role_name: this.userRole
+        enabled: this.enabled,
+        default_device_connectivity: this.defaultConnectivity,
+        organization_id: this.organizationId,
+        device_group_api_key: this.groupApiKey
       };
-      this.backend.addUser(this.organizationId, user)
-        .subscribe(response => {
-          this.notificationsService.add({
-            message: 'The user ' + user.email + ' has been created successfully'
-          });
-          this.bsModalRef.hide();
-        }, error => {
-          if (error.status === 409) {
-            this.notificationsService.add({
-              message: 'ERROR: ' + error.error.message + ' already exists',
-              timeout: 10000
-            });
-          } else {
-              this.notificationsService.add({
-                message: 'ERROR: ' + error.error.message,
-                timeout: 10000
-              });
-            }
+      this.backend.addGroup(this.organizationId, groupData)
+      .subscribe(response => {
+        const initialState = {
+          groupApiKey: response.device_group_api_key,
+        };
+        this.bsModalRef.hide();
+        this.bsModalRef =
+          this.modalService.show(DeviceGroupCreatedComponent, { initialState, backdrop: 'static', ignoreBackdropClick: false });
+        this.bsModalRef.content.closeBtnName = 'Close';
+      }, error => {
+        this.notificationsService.add({
+          message: 'ERROR: ' + error.error.message,
+          timeout: 10000
         });
+      });
     }
   }
 
@@ -119,101 +121,4 @@ export class AddDevicesGroupComponent implements OnInit {
     }
   }
 
-  /**
-   * Changes the new user role
-   * @param newRole New user role
-   */
-  changeRole(newRole) {
-    this.userRole = newRole;
-  }
-
-  /**
-   * Validates user data
-   * @param form Form with user data
-   */
-  checkFormFields(form: FormGroup) {
-    this.errorMessages = [];
-    if (form.controls.email.invalid) {
-      if (form.controls.email.errors.required) {
-        this.errorMessages.push('Email is required');
-      }
-      if (form.controls.email.errors.pattern) {
-        this.errorMessages.push('Email must be a valid email address');
-      }
-    }
-    if (form.controls.password.invalid) {
-      if (form.controls.password.errors.required) {
-        this.errorMessages.push('Password is required');
-      }
-      if (form.controls.password.errors.minlength) {
-        this.errorMessages.push('Password must have more than 6 characters');
-      }
-    }
-    if (form.controls.password.value !== form.controls.passwordConfirm.value) {
-      this.errorMessages.push('Password and confirm password are not the same one');
-    }
-    if (!this.userRole) {
-      this.errorMessages.push('User role selection is required');
-    }
-
-    if (this.errorMessages.length === 0) {
-      this.addUser(form);
-    }
-  }
-
-  /**
-   * Outputs the error messages in the required format, showing the first one
-   * @param errors String containing the errors
-   */
-  formatValidationOutput(errors: string[]) {
-    if (this.errorMessages.length === 1) {
-      return {
-        msg: this.errorMessages[0],
-        errors: this.errorMessages
-      };
-    } else if (this.errorMessages.length > 0) {
-      return {
-        msg: this.errorMessages[0] + ' +' + (this.errorMessages.length - 1) + ' errors',
-        errors: this.errorMessages
-      };
-    } else {
-      return {
-        msg: '',
-        errors: this.errorMessages
-      };
-    }
-  }
-
-  /**
-   * Another string definition of an array
-   * @param array Array of elements
-   */
-  arrayToString(array: any[]): string {
-    let msg = '';
-    array.forEach(element => {
-      msg = msg + element.toLowerCase() + ', ';
-    });
-    msg = msg.slice(0, msg.length - 2);
-    return msg;
-  }
-
-  /**
-   * Checkbox
-   */
-  enabled() {
-    this.checkBox = !this.checkBox;
-  }
-
-  /**
-   * Create device group opens the device group created component modal window confirmation
-   */
-  createDeviceGroup() {
-    const initialState = {
-      organizationId: this.organizationId,
-    };
-    this.bsDeviceGroupModalRef =
-      this.modalService.show(DeviceGroupCreatedComponent, { initialState, backdrop: 'static', ignoreBackdropClick: false });
-    this.bsDeviceGroupModalRef.content.closeBtnName = 'Close';
-    this.bsModalRef.hide();
-  }
 }