+++ src/app/utils/mocks.ts
@@ -1,3 +1,5 @@
+import { Group } from '../definitions/interfaces/group';
+
 export const mockJwtToken = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9' +
     // tslint:disable-next-line:max-line-length
     '.eyJhY2Nlc3MiOlsiT1JHIl0sImV4cCI6MTU0MjI4Njg2MywiaWF0IjoxNTQyMjc2MDYzLCJpc3MiOiJhdXRoeCIsImp0aSI6IjI1OTA5ZDNkLTJlODMtNDlmMC04ZmQzLTFlYmZiNTYxMTNhMSIsIm5iZiI6IjE1NDIyNzYwNjMiLCJvcmdhbml6YXRpb25JRCI6IjdhZDFhN2E4LWU0YjEtNDc5OC05MDcxLWU0NTY5MDhmYWQxMyIsInJvbGUiOiJPd25lciIsInVzZXJJRCI6ImpvaG4uZG9lQG1haWwuY29tIn0' +
@@ -1584,3 +1586,254 @@ export const mockClusterDetail = [
         multitenant_support: 'yes'
     }
 ];
+
+
+/**
+ * Mocked Devices status timeline
+ */
+export const mockDevicesChart = [
+    {
+        'name': 'Devices Running',
+        'series': [
+          {
+            'value': 39,
+            'name': '-6h'
+          },
+          {
+            'value': 79,
+            'name': '-5h'
+          },
+          {
+            'value': 23,
+            'name': '-4h'
+          },
+          {
+            'value': 96,
+            'name': '-3h'
+          },
+          {
+            'value': 40,
+            'name': '-2h'
+          },
+          {
+            'value': 73,
+            'name': '-1h'
+          },
+          {
+            'value': 23,
+            'name': 'now'
+          }
+        ]
+      }
+];
+
+/**
+ * Mocked devices list
+ */
+export const mockDevicesList = [
+    [
+        {
+          organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+          device_group_id: 'a1',
+          device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+          register_since: 1550506924515,
+          labels: {
+              type: 'phone',
+              os: 'arm',
+          },
+          enabled: 'true',
+          device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+          status_name: 'Connected'
+        },
+        {
+            organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+            device_group_id: 'a1',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1550417971000,
+            labels: {
+                type: 'phone',
+            },
+            enabled: 'true',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'Connected'
+        }
+    ],
+    [
+        {
+            organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+            device_group_id: 'b2',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1549899571000,
+            labels: {
+                type: 'phone',
+                os: 'arm',
+            },
+            enabled: 'true',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'Connected'
+        },
+        {
+            organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+            device_group_id: 'b2',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1547825971000,
+            labels: {
+                type: 'phone',
+            },
+            enabled: 'true',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'Connected'
+        },
+        {
+            organization_id: '7ad1a7a8-e4b1-4798-9071-e456908fad13',
+            device_group_id: 'b2',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1518968371000,
+            labels: {
+                type: 'phone',
+                os: 'arm',
+            },
+            enabled: 'false',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'disconnected'
+        },
+        {
+            organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+            device_group_id: 'b2',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1516087991000,
+            labels: {
+                type: 'phone',
+            },
+            enabled: 'true',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'Connected'
+        }
+    ],
+    [
+        {
+            organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+            device_group_id: 'c3',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1527261837000,
+            labels: {
+                type: 'phone',
+                os: 'arm',
+            },
+            enabled: 'true',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'disconnected'
+        },
+        {
+            organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+            device_group_id: 'c3',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1515486791000,
+            labels: {
+                type: 'phone',
+            },
+            enabled: 'true',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'disconnected'
+        },
+        {
+            organization_id: '7ad1a7a8-e4b1-4798-9071-e456908fad13',
+            device_group_id: 'c3',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1513413191000,
+            labels: {
+                type: 'phone',
+                os: 'arm',
+            },
+            enabled: 'false',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'disconnected'
+        },
+        {
+            organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+            device_group_id: 'c3',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1527261837000,
+            labels: {
+                type: 'phone',
+            },
+            enabled: 'true',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'disconnected'
+        }
+    ],
+    [
+        {
+            organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+            device_group_id: 'd4',
+            device_id: '792989c-4ae4-460f-92b5-bca7ed36f016',
+            register_since: 1497108201000,
+            labels: {
+                type: 'phone',
+                os: 'arm',
+            },
+            enabled: 'true',
+            device_api_key: '7bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3',
+            status_name: 'disconnected'
+        }
+    ]
+];
+
+/**
+ * Mocked devices group list
+ */
+export const mockGroupList: Group[] = [
+    {
+        organization_id: 'b792989c-4ae4-460f-92b5-bca7ed36f016',
+        device_group_id: 'a1',
+        enabled: true,
+        default_device_connectivity: true,
+        name: 'Voice controllers',
+        device_group_api_key: '2bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3c11ac797c79'
+    },
+    {
+        organization_id: 'a792989c-4ae4-460f-92b5-bca7ed36f017',
+        device_group_id: 'b2',
+        enabled: false,
+        default_device_connectivity: false,
+        name: 'Nest cams Outdoor',
+        device_group_api_key: '3bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3c11ac797c78'
+    },
+    {
+        organization_id: 'a792989c-4ae4-460f-92b5-bca7ed36f017',
+        device_group_id: 'c3',
+        enabled: true,
+        default_device_connectivity: true,
+        name: 'Tracking devices',
+        device_group_api_key: '4bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3c11ac797c76'
+    },
+    {
+        organization_id: 'a792989c-4ae4-460f-92b5-bca7ed36f017',
+        device_group_id: 'd4',
+        enabled: true,
+        default_device_connectivity: true,
+        name: 'Bluetooth sensors',
+        device_group_api_key: '5bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3c11ac797c75'
+    },
+    {
+        organization_id: 'a792989c-4ae4-460f-92b5-bca7ed36f017',
+        device_group_id: 'e5',
+        enabled: true,
+        default_device_connectivity: true,
+        name: 'Droids with force',
+        device_group_api_key: '5bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3c11ac797c75'
+    },
+    {
+        organization_id: 'a792989c-4ae4-460f-92b5-bca7ed36f017',
+        device_group_id: 'f6',
+        enabled: true,
+        default_device_connectivity: true,
+        name: 'Smart thermostats',
+        device_group_api_key: '5bd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-1008-ac7a-398a-3c11ac797c75'
+    }
+  ];
+
+/**
+ * Mocked new group API Key
+ */
+export const mockGroupApiKey = '9cd7d59cfe90e4d32b1d2f20d39c86df-fbaa8670-3352-ac7a-398a-3c11ac797c74';